add Command
