+++ build.gradle
@@ -53,7 +53,7 @@ test {
 }
 
 application {
-    mainClass.set("seedu.duke.Duke")
+    mainClass.set("seedu.duke.Main")
 }
 
 shadowJar {
