--- src/main/java/seedu/duke/data/duke.txt
@@ -1,7 +1,0 @@
-[T][X] homework
-[D][ ] fish (by: 2000-12-20)
-[T][ ] homework
-[D][ ] cook (by: 2022-12-20)
-[T][X] homework
-[D][ ] hw2 (by: 2012-01-12)
-[E][ ] party (from: monday to: tuesday)
