--- src/main/java/seedu/duke/package.java
@@ -1,7 +1,0 @@
-package duke;
-
-class Duke {
-    public static void main(String[] args) {
-
-    }
-}
